# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Simple open source polling REST API.",
        "title": "Poll API",
        "contact": {},
        "version": "1.0"
    },
    "host": "https://poll-api.vncsmyrnk.dev",
    "basePath": "/api",
    "paths": {
        "/auth/logout": {
            "post": {
                "description": "Clears the refresh token cookie",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logs the autheticated user out",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "Creates a new access token cookie based on the refresh token. This cookie is used as authentication for `/api` calls.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refreshes the autheticated user out",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/polls": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "polls"
                ],
                "summary": "Lists all polls available",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "current results page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "name search by q",
                        "name": "q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Poll"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "polls"
                ],
                "summary": "Creates a new poll",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ports.CreatePollInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/polls/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "polls"
                ],
                "summary": "Get a poll by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "poll id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Poll"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/polls/{id}/count": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "polls"
                ],
                "summary": "Get the current vote count for a poll",
                "parameters": [
                    {
                        "type": "string",
                        "description": "authorization header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "poll id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/domain.PollOptionStats"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/polls/{id}/my-vote": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "polls"
                ],
                "summary": "Gets the user vote on a poll",
                "parameters": [
                    {
                        "type": "string",
                        "description": "authorization header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "poll id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Vote"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/polls/{id}/votes": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "polls"
                ],
                "summary": "Casts a vote on a poll",
                "parameters": [
                    {
                        "type": "string",
                        "description": "authorization header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.voteRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "poll id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Poll": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.PollOption"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "domain.PollOption": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "poll_id": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "domain.PollOptionStats": {
            "type": "object",
            "properties": {
                "percentage": {
                    "type": "number",
                    "format": "float64"
                },
                "voteCount": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
        "domain.Vote": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "option_id": {
                    "type": "string"
                },
                "poll_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "voter_ip": {
                    "type": "string"
                }
            }
        },
        "http.voteRequest": {
            "type": "object",
            "properties": {
                "option_id": {
                    "type": "string"
                }
            }
        },
        "ports.CreatePollInput": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}