# internal/adapters/handler/http/poll_handler.go
package http

import (
	"encoding/json"
	"errors"
	"net/http"
	"strconv"

	"github.com/go-chi/chi/v5"
	"github.com/google/uuid"
	"github.com/vncsmyrnk/poll/internal/core/domain"
	"github.com/vncsmyrnk/poll/internal/core/ports"
)

type PollHandler struct {
	service ports.PollService
}

func NewPollHandler(service ports.PollService) *PollHandler {
	return &PollHandler{
		service: service,
	}
}

type createPollRequest struct {
	Title       string   `json:"title"`
	Description string   `json:"description"`
	Options     []string `json:"options"`
}

// CreatePoll godoc
// @Summary      Creates a new poll
// @Tags         polls
// @Accept       json
// @Param        request body ports.CreatePollInput true "request body"
// @Success      201
// @Failure      500
// @Router       /polls [post]
func (h *PollHandler) CreatePoll(w http.ResponseWriter, r *http.Request) {
	var req createPollRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "invalid request body", http.StatusBadRequest)
		return
	}

	input := ports.CreatePollInput{
		Title:       req.Title,
		Description: req.Description,
		Options:     req.Options,
	}

	poll, err := h.service.Create(r.Context(), input)
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusCreated)
	if err := json.NewEncoder(w).Encode(poll); err != nil {
		http.Error(w, "failed to encode response", http.StatusInternalServerError)
	}
}

// GetPoll godoc
// @Summary      Get a poll by id
// @Tags         polls
// @Produce      json
// @Param        id   path      int  true  "poll id"
// @Success      200  {object}  domain.Poll
// @Failure      400
// @Failure      404
// @Failure      500
// @Router       /polls/{id} [get]
func (h *PollHandler) GetPoll(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")
	if id == "" {
		http.Error(w, "missing poll id", http.StatusBadRequest)
		return
	}

	poll, err := h.service.GetPoll(r.Context(), id)
	if err != nil {
		if errors.Is(err, domain.ErrInvalidPollID) {
			http.Error(w, err.Error(), http.StatusBadRequest)
			return
		}
		if errors.Is(err, domain.ErrPollNotFound) {
			http.Error(w, err.Error(), http.StatusNotFound)
			return
		}

		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	if err := json.NewEncoder(w).Encode(poll); err != nil {
		http.Error(w, "failed to encode response", http.StatusInternalServerError)
	}
}

// ListPolls godoc
// @Summary      Lists all polls available
// @Tags         polls
// @Produce      json
// @Param        page   query      int  fale  "current results page"
// @Param        q      query   string  false "name search by q"
// @Success      200  {object}  []domain.Poll
// @Failure      500
// @Router       /polls [get]
func (h *PollHandler) ListPolls(w http.ResponseWriter, r *http.Request) {
	pageStr := r.URL.Query().Get("page")
	page := 1
	if pageStr != "" {
		p, err := strconv.Atoi(pageStr)
		if err == nil && p > 0 {
			page = p
		}
	}

	query := r.URL.Query().Get("q")

	input := ports.ListPollsInput{
		Page:  page,
		Query: query,
	}

	polls, err := h.service.ListPolls(r.Context(), input)
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	if err := json.NewEncoder(w).Encode(polls); err != nil {
		http.Error(w, "failed to encode response", http.StatusInternalServerError)
	}
}

// GetPollStats godoc
// @Summary      Get the current vote count for a poll
// @Tags         polls
// @Produce      json
// @Param        Authorization    header    string    true   	"authorization header"
// @Param        id   path      int  true  "poll id"
// @Success      200  {object}  map[uuid.UUID]domain.PollOptionStats
// @Failure      400
// @Failure      403
// @Failure      404
// @Failure      500
// @Router       /polls/{id}/count [get]
func (h *PollHandler) GetPollStats(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")
	if id == "" {
		http.Error(w, "missing poll id", http.StatusBadRequest)
		return
	}

	userID, ok := r.Context().Value(UserIDKey).(uuid.UUID)
	if !ok {
		http.Error(w, "Unauthorized: missing user context", http.StatusUnauthorized)
		return
	}

	stats, err := h.service.GetPollStats(r.Context(), id, userID)
	if err != nil {
		if errors.Is(err, domain.ErrInvalidPollID) {
			http.Error(w, err.Error(), http.StatusBadRequest)
			return
		}
		if errors.Is(err, domain.ErrPollNotFound) {
			http.Error(w, err.Error(), http.StatusNotFound)
			return
		}
		if errors.Is(err, domain.ErrUserNotVoted) {
			http.Error(w, err.Error(), http.StatusForbidden)
			return
		}

		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	if err := json.NewEncoder(w).Encode(stats); err != nil {
		http.Error(w, "failed to encode response", http.StatusInternalServerError)
	}
}